Simplifies EVERYTHING
The original version of this was designed to generate events that
actually made sense. For example, 3 events on the same card that all had
the same uuid on the card, and the dates were chronological as far as
the created date, update, deleted, etc.

The intent was to make as acurate data as possible for testing
elasticsearch performance. Elasticsearch is plenty fast enough after
testing.

Now I have changed the json schema to a flatter design, which should
make searching even faster. But changing this code to rearrange the
structure of the generated json was over complicated.

This refactoring deletes much of the logic behind making data that makes
sense, and now only generates one off events. There shouldn't ever be 2
events with the same card uuid (unless there is a uuid collision). So
technically none of the data would actually be bad, there's just no
"series" of events any more.